check if grizzly-loadtester is installed with any extras
return grizzly version from get_dependency_versions as a tuple of a
string and an optional list of strings, where the list is empty if
no extras, or a list of the extras specified in requirements.txt.

if there are extras, show them next to the grizzly-loadtester version in
`--version all`.