no smoothed timeline in test summary
in biometria-se/grizzly#84 the behavior is changed to used fixed_count
for a user (based on user weight and total number of users).

this means that the timeline is not applicable anymore, so instead
calculate how many users a scenario will get from the "pool", and show
it in the summary.