keyvault improvements (#109)
* check for empty configuration file

easiest mistake to reproduce would be to misspell a key filter (`-k`).

copy original environment file for import also, before rewriting the specified file with imported configuration.

* show all files that was written for mqm certificates, not only the kdb file.

* improvements

fixed that `-k` filters on a partial path name to a configuration key (e.g. startswith).

fixed that, when merging dicts, if source is None, but merged is a dict, do not overwrite that with the None value, just keep it as it is.